successfully completed update and delete api

+++ biometric_ticket/views.py
@@ -1,11 +1,13 @@
 from django.shortcuts import render
 
-from django.http import HttpResponse
+from django.http import HttpResponse, Http404
+
+from django.core.exceptions import ObjectDoesNotExist
 
 from rest_framework import generics, permissions, status
 from rest_framework.parsers import JSONParser
 from .models import LED_bulb
-
+from rest_framework.decorators import api_view
 
 
 # for LED Buld only
@@ -35,14 +37,66 @@ class LED_bulbList(APIView):
             serializer.save()
             return Response(serializer.data, status = status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+class LED_bulb_pd(APIView):
+    
+    def get_object(self, id):
+        try:
+            return LED_bulb.objects.get(id = id)
+        except ObjectDoesNotExist:
+            raise Http404
+
+    def get(self, request, id, format = None):
+        led_bulb = self.get_object(id)
+        serializer = LED_bulbSerializer(led_bulb)
+        return Response(serializer.data)
+
     def put(self, request, id, format = None):
-        led_bulb = LED_bulb.objects.get(id = id)
+        led_bulb =self.get_object(id)
         serializer = LED_bulbSerializer(led_bulb, data = request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status = status.HTTP_201_CREATED)
+        # print(serializer.errors)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     def delete(self, request, id, format = None):
+        led_bulb = self.get_object(id)
+        led_bulb.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
+@api_view(['GET', 'POST'])
+def led_bulb_list(request, format = None):
+    if request.method == 'GET':
+        led_bulb = LED_bulb.objects.all()
+        serializer = LED_bulbSerializer(led_bulb, many = True)
+        return Response(serializer.data)
+    elif request.method == 'POST':
+        # data = JSONParser.parse(request.data)
+        data = request.data
+        serializer = LED_bulbSerializer(data = data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status = status.HTTP_400_BAD_REQUEST)
+
+@api_view(['GET', 'PUT', 'DELETE'])
+def led_bulb_detail(request, id, format = None):
+    try:
         led_bulb = LED_bulb.objects.get(id = id)
+    except led_bulb.DeosNotExist:
+        return Response(status=status.HTTP_404_NOT_FOUND)
+    
+    if request.method == 'GET':
+        serializer = LED_bulbSerializer(led_bulb)
+        return Response(serializer.data)
+
+    elif request.method == 'PUT':
+        serializer = LED_bulbSerializer(led_bulb, data = request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    elif request.method == 'DELETE':
         led_bulb.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
